Editors are moved to separate files
